Updating exceptions

+++ clinica-odontologica/src/main/java/com/dh/clinica/service/OdontologoServiceImpl.java
@@ -1,5 +1,6 @@
 package com.dh.clinica.service;
 
+import com.dh.clinica.exceptions.ResourceNotFoundException;
 import com.dh.clinica.model.Odontologo;
 import com.dh.clinica.repository.OdontologoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,7 +35,10 @@ public class OdontologoServiceImpl implements OdontologoService {
     }
 
     @Override
-    public void eliminar(Integer id) {
+    public void eliminar(Integer id) throws ResourceNotFoundException {
+        if(obtener(id).isEmpty())
+            throw  new ResourceNotFoundException("No existe odontologo con id: "+id);
+
         odontologoRepository.deleteById(id);
     }
