+++ clinica-odontologica/src/main/java/com/dh/clinica/service/TurnoServiceImpl.java
@@ -1,6 +1,7 @@
 package com.dh.clinica.service;
 
 import com.dh.clinica.exceptions.BadRequestException;
+import com.dh.clinica.exceptions.ResourceNotFoundException;
 import com.dh.clinica.model.Turno;
 import com.dh.clinica.repository.TurnoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,10 +13,16 @@ import java.util.Optional;
 @Service
 public class TurnoServiceImpl implements TurnoService {
     private final TurnoRepository turnoRepository;
+    private PacienteService pacienteService;
+    private OdontologoService odontologoService;
 
 
     @Autowired
-    public TurnoServiceImpl(TurnoRepository turnoRepository){this.turnoRepository = turnoRepository;}
+    public TurnoServiceImpl(TurnoRepository turnoRepository, PacienteService pacienteService, OdontologoService odontologoService) {
+        this.turnoRepository = turnoRepository;
+        this.pacienteService = pacienteService;
+        this.odontologoService = odontologoService;
+    }
 
     @Override
     public List<Turno> obtenerTodos() {
@@ -29,14 +36,16 @@ public class TurnoServiceImpl implements TurnoService {
 
     @Override
     public Turno guardar(Turno turno) throws BadRequestException{
-        if(turno.getOdontologo().getId() ==null || turno.getPaciente().getId()==null){
+        if(pacienteService.obtener(turno.getPaciente().getId()).isEmpty() || odontologoService.obtener(turno.getOdontologo().getId()).isEmpty()){
             throw new BadRequestException("No existe el odontologo/ paciente requerido");
         }
         return turnoRepository.save(turno);
     }
 
     @Override
-    public void eliminar(Integer id) {
+    public void eliminar(Integer id) throws ResourceNotFoundException {
+        if(obtener(id).isEmpty())
+            throw  new ResourceNotFoundException("No existe turno con id: "+id);
         turnoRepository.deleteById(id);
     }
